fix: implement native playwright sandbox architecture for UI captures
